+++ app/system/modules/user/views/reset-request.php
@@ -1,18 +1,15 @@
-<form class="uk-article uk-form uk-form-stacked" action="<?= $view->url('@user/resetpassword/request') ?>" method="post">
+<form class="uk-form uk-form-stacked uk-width-medium-1-2 uk-width-large-1-3 uk-container-center" action="<?= $view->url('@user/resetpassword/request') ?>" method="post">
 
-    <h1 class="uk-article-title"><?= __('Forgot Password') ?></h1>
+    <h1 class="uk-h2 uk-text-center"><?= __('Forgot Password') ?></h1>
 
     <p><?= __('Please enter your email address. You will receive a link to create a new password via email.') ?></p>
 
     <div class="uk-form-row">
-        <label for="form-username" class="uk-form-label"><?= __('Email') ?></label>
-        <div class="uk-form-controls">
-            <input id="form-username" class="uk-form-width-medium" type="text" name="email" value="" required autofocus>
-        </div>
+        <input class="uk-width-1-1" type="text" name="email" value="" placeholder="<?= __('Email') ?>" required autofocus>
     </div>
 
     <p class="uk-form-row">
-        <button class="uk-button uk-button-primary" type="submit"><?= __('Submit') ?></button>
+        <button class="uk-button uk-button-primary uk-button-large uk-width-1-1" type="submit"><?= __('Request password') ?></button>
     </p>
 
     <?php $view->token()->get() ?>